fix: Array type would `append_path` instead of `add_path` on error
The error path during array failure was wrong.
All other types seems to call `add_path`
(which should probably be called `prepend_path`)
on it's way out of nested parsing.

Array was the only one using `append_path`, so I've removed `append_path`.

I also fixed a test that asserted the wrong path.
This is what gave it away, the path in the test was not pointing to the
actual error element of the data.